Compilation fix for video
git-svn-id: svn://svn.minisip.org/minisip/trunk@1565 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ minisip/video/codec/VideoCodec.cxx
@@ -54,7 +54,7 @@ std::string VideoCodec::getCodecDescription(){
 	return std::string( "ITU-T H.263 Video Encoder/Decoder" );
 }
 
-int32_t VideoCodec::getSdpMediaType(){
+uint8_t VideoCodec::getSdpMediaType(){
 	return 105;
 }
